imtgeo/CTileProvider.cpp
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@10886 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtgeo/CMapTileCollectionControllerComp.cpp
@@ -75,7 +75,7 @@ imtbase::CTreeItemModel* CMapTileCollectionControllerComp::ListObjects(const imt
 		double rightBottomLat = 0;
 		double rightBottomLon = 0;
 
-		QList<QPair<QPair<int, int>, int>> tileList;
+		QList<QPair<QPair<qint64, qint64>, int>> tileList;
 
 		if (viewParamsGql != nullptr){
 			QByteArray filterBA = viewParamsGql->GetFieldArgumentValue("FilterModel").toByteArray();
